[FIX] base: move _load_module_terms to ir.module.module
